IGNITE-15052 Moves shared thread pools management to PoolProcessor.
