+++ modules/core/src/main/java/org/apache/ignite/internal/GridKernalContext.java
@@ -21,7 +21,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.Executor;
-import java.util.concurrent.ExecutorService;
 import org.apache.ignite.IgniteLogger;
 import org.apache.ignite.configuration.IgniteConfiguration;
 import org.apache.ignite.internal.cache.query.index.IndexProcessor;
@@ -78,14 +77,11 @@ import org.apache.ignite.internal.processors.timeout.GridTimeoutProcessor;
 import org.apache.ignite.internal.processors.tracing.Tracing;
 import org.apache.ignite.internal.suggestions.GridPerformanceSuggestions;
 import org.apache.ignite.internal.util.IgniteExceptionRegistry;
-import org.apache.ignite.internal.util.StripedExecutor;
 import org.apache.ignite.internal.util.tostring.GridToStringExclude;
 import org.apache.ignite.internal.worker.WorkersRegistry;
 import org.apache.ignite.maintenance.MaintenanceRegistry;
 import org.apache.ignite.plugin.PluginNotFoundException;
 import org.apache.ignite.plugin.PluginProvider;
-import org.apache.ignite.thread.IgniteStripedThreadPoolExecutor;
-import org.jetbrains.annotations.Nullable;
 
 /**
  *
@@ -331,20 +327,6 @@ public interface GridKernalContext extends Iterable<GridComponent> {
     public GridMarshallerMappingProcessor mapping();
 
     /**
-     * Gets utility cache pool.
-     *
-     * @return Utility cache pool.
-     */
-    public ExecutorService utilityCachePool();
-
-    /**
-     * Gets async callback pool.
-     *
-     * @return Async callback pool.
-     */
-    public IgniteStripedThreadPoolExecutor asyncCallbackPool();
-
-    /**
      * Gets cache object processor.
      *
      * @return Cache object processor.
@@ -555,115 +537,6 @@ public interface GridKernalContext extends Iterable<GridComponent> {
     public <T> T createComponent(Class<T> cls);
 
     /**
-     * @return Thread pool implementation to be used in grid to process job execution
-     *      requests and user messages sent to the node.
-     */
-    public ExecutorService getExecutorService();
-
-    /**
-     * Executor service that is in charge of processing service proxy invocations.
-     *
-     * @return Thread pool implementation to be used in grid for service proxy invocations.
-     */
-    public ExecutorService getServiceExecutorService();
-
-    /**
-     * Executor service that is in charge of processing internal system messages.
-     *
-     * @return Thread pool implementation to be used in grid for internal system messages.
-     */
-    public ExecutorService getSystemExecutorService();
-
-    /**
-     * Executor service that is in charge of processing internal system messages
-     * in stripes (dedicated threads).
-     *
-     * @return Thread pool implementation to be used in grid for internal system messages.
-     */
-    public StripedExecutor getStripedExecutorService();
-
-    /**
-     * Executor service that is in charge of processing internal and Visor
-     * {@link org.apache.ignite.compute.ComputeJob GridJobs}.
-     *
-     * @return Thread pool implementation to be used in grid for internal and Visor
-     *      jobs processing.
-     */
-    public ExecutorService getManagementExecutorService();
-
-    /**
-     * @return Thread pool implementation to be used for peer class loading
-     *      requests handling.
-     */
-    public ExecutorService getPeerClassLoadingExecutorService();
-
-    /**
-     * Executor service that is in charge of processing data stream messages.
-     *
-     * @return Thread pool implementation to be used for data stream messages.
-     */
-    public StripedExecutor getDataStreamerExecutorService();
-
-    /**
-     * Should return an instance of fully configured thread pool to be used for
-     * processing of client messages (REST requests).
-     *
-     * @return Thread pool implementation to be used for processing of client
-     *      messages.
-     */
-    public ExecutorService getRestExecutorService();
-
-    /**
-     * Get affinity executor service.
-     *
-     * @return Affinity executor service.
-     */
-    public ExecutorService getAffinityExecutorService();
-
-    /**
-     * Get indexing executor service.
-     *
-     * @return Indexing executor service.
-     */
-    @Nullable public ExecutorService getIndexingExecutorService();
-
-    /**
-     * Executor service that is in charge of processing query messages.
-     *
-     * @return Thread pool implementation to be used in grid for query messages.
-     */
-    public ExecutorService getQueryExecutorService();
-
-
-    /**
-     * Executor services that is in charge of processing user compute task.
-     *
-     * @return Map of custom thread pool executors.
-     */
-    @Nullable public Map<String, ? extends ExecutorService> customExecutors();
-
-    /**
-     * Executor service that is in charge of processing schema change messages.
-     *
-     * @return Executor service that is in charge of processing schema change messages.
-     */
-    public ExecutorService getSchemaExecutorService();
-
-    /**
-     * Executor service that is in charge of processing rebalance messages.
-     *
-     * @return Executor service that is in charge of processing rebalance messages.
-     */
-    public ExecutorService getRebalanceExecutorService();
-
-    /**
-     * Executor service that is in charge of processing unorderable rebalance messages.
-     *
-     * @return Executor service that is in charge of processing unorderable rebalance messages.
-     */
-    public IgniteStripedThreadPoolExecutor getStripedRebalanceExecutorService();
-
-    /**
      * Gets exception registry.
      *
      * @return Exception registry.
@@ -763,13 +636,6 @@ public interface GridKernalContext extends Iterable<GridComponent> {
      * @return Local continuous tasks processor.
      */
     public DurableBackgroundTasksProcessor durableBackgroundTask();
-
-    /**
-     * Return Thread pool for create/rebuild indexes.
-     *
-     * @return Thread pool for create/rebuild indexes.
-     */
-    public ExecutorService buildIndexExecutorService();
 
     /**
      * Gets Performance statistics processor.
