+++ modules/core/src/main/java/org/apache/ignite/internal/IgnitionEx.java
@@ -38,9 +38,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.Executors;
-import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.logging.Handler;
@@ -75,7 +73,6 @@ import org.apache.ignite.configuration.TransactionConfiguration;
 import org.apache.ignite.failure.FailureContext;
 import org.apache.ignite.failure.FailureType;
 import org.apache.ignite.internal.binary.BinaryMarshaller;
-import org.apache.ignite.internal.managers.communication.GridIoPolicy;
 import org.apache.ignite.internal.managers.discovery.GridDiscoveryManager;
 import org.apache.ignite.internal.processors.cache.CacheGroupContext;
 import org.apache.ignite.internal.processors.cache.distributed.dht.preloader.GridDhtPartitionFullMap;
@@ -89,7 +86,6 @@ import org.apache.ignite.internal.processors.resource.DependencyResolver;
 import org.apache.ignite.internal.processors.resource.GridSpringResourceContext;
 import org.apache.ignite.internal.util.GridConcurrentHashSet;
 import org.apache.ignite.internal.util.IgniteUtils;
-import org.apache.ignite.internal.util.StripedExecutor;
 import org.apache.ignite.internal.util.TimeBag;
 import org.apache.ignite.internal.util.spring.IgniteSpringHelper;
 import org.apache.ignite.internal.util.typedef.CA;
@@ -106,7 +102,6 @@ import org.apache.ignite.internal.util.worker.GridWorker;
 import org.apache.ignite.internal.worker.WorkersRegistry;
 import org.apache.ignite.lang.IgniteBiInClosure;
 import org.apache.ignite.lang.IgniteBiTuple;
-import org.apache.ignite.lang.IgniteInClosure;
 import org.apache.ignite.marshaller.Marshaller;
 import org.apache.ignite.marshaller.MarshallerUtils;
 import org.apache.ignite.marshaller.jdk.JdkMarshaller;
@@ -129,9 +124,7 @@ import org.apache.ignite.spi.loadbalancing.LoadBalancingSpi;
 import org.apache.ignite.spi.loadbalancing.roundrobin.RoundRobinLoadBalancingSpi;
 import org.apache.ignite.spi.metric.noop.NoopMetricExporterSpi;
 import org.apache.ignite.spi.tracing.NoopTracingSpi;
-import org.apache.ignite.thread.IgniteStripedThreadPoolExecutor;
 import org.apache.ignite.thread.IgniteThread;
-import org.apache.ignite.thread.IgniteThreadPoolExecutor;
 import org.jetbrains.annotations.Nullable;
 
 import static java.util.stream.Collectors.joining;
@@ -153,10 +146,8 @@ import static org.apache.ignite.cache.CacheMode.PARTITIONED;
 import static org.apache.ignite.cache.CacheMode.REPLICATED;
 import static org.apache.ignite.cache.CacheRebalanceMode.SYNC;
 import static org.apache.ignite.cache.CacheWriteSynchronizationMode.FULL_SYNC;
-import static org.apache.ignite.configuration.IgniteConfiguration.DFLT_THREAD_KEEP_ALIVE_TIME;
 import static org.apache.ignite.configuration.MemoryConfiguration.DFLT_MEMORY_POLICY_MAX_SIZE;
 import static org.apache.ignite.configuration.MemoryConfiguration.DFLT_MEM_PLC_DEFAULT_NAME;
-import static org.apache.ignite.failure.FailureType.SYSTEM_WORKER_TERMINATION;
 import static org.apache.ignite.internal.IgniteComponentType.SPRING;
 import static org.apache.ignite.plugin.segmentation.SegmentationPolicy.RESTART_JVM;
 
@@ -1608,60 +1599,6 @@ public class IgnitionEx {
         /** Grid instance. */
         private volatile IgniteKernal grid;
 
-        /** Executor service. */
-        private ThreadPoolExecutor execSvc;
-
-        /** Executor service for services. */
-        private ThreadPoolExecutor svcExecSvc;
-
-        /** System executor service. */
-        private ThreadPoolExecutor sysExecSvc;
-
-        /** */
-        private StripedExecutor stripedExecSvc;
-
-        /** Management executor service. */
-        private ThreadPoolExecutor mgmtExecSvc;
-
-        /** P2P executor service. */
-        private ThreadPoolExecutor p2pExecSvc;
-
-        /** Data streamer executor service. */
-        private StripedExecutor dataStreamerExecSvc;
-
-        /** REST requests executor service. */
-        private ThreadPoolExecutor restExecSvc;
-
-        /** Utility cache executor service. */
-        private ThreadPoolExecutor utilityCacheExecSvc;
-
-        /** Affinity executor service. */
-        private ThreadPoolExecutor affExecSvc;
-
-        /** Indexing pool. */
-        private ThreadPoolExecutor idxExecSvc;
-
-        /** Thread pool for create/rebuild indexes. */
-        private ThreadPoolExecutor buildIdxExecSvc;
-
-        /** Continuous query executor service. */
-        private IgniteStripedThreadPoolExecutor callbackExecSvc;
-
-        /** Query executor service. */
-        private ThreadPoolExecutor qryExecSvc;
-
-        /** Query executor service. */
-        private ThreadPoolExecutor schemaExecSvc;
-
-        /** Rebalance executor service. */
-        private ThreadPoolExecutor rebalanceExecSvc;
-
-        /** Rebalance striped executor service. */
-        private IgniteStripedThreadPoolExecutor rebalanceStripedExecSvc;
-
-        /** Executor service. */
-        private Map<String, ThreadPoolExecutor> customExecSvcs;
-
         /** Grid state. */
         private volatile IgniteState state = STOPPED;
 
@@ -1786,12 +1723,6 @@ public class IgnitionEx {
                         log.info("Node started : "
                             + startNodeTimer.stagesTimings().stream().collect(joining(",", "[", "]")));
                 }
-                catch (Exception e) {
-                    if (log != null)
-                        stopExecutors(log);
-
-                    throw e;
-                }
                 finally {
                     startLatch.countDown();
                 }
@@ -1824,63 +1755,12 @@ public class IgnitionEx {
                 ensureMultiInstanceSupport(cfg.getLoadBalancingSpi());
             }
 
-            validateThreadPoolSize(cfg.getPublicThreadPoolSize(), "public");
-
             UncaughtExceptionHandler oomeHnd = new UncaughtExceptionHandler() {
                 @Override public void uncaughtException(Thread t, Throwable e) {
                     if (grid != null && X.hasCause(e, OutOfMemoryError.class))
                         grid.context().failure().process(new FailureContext(FailureType.CRITICAL_ERROR, e));
                 }
             };
-
-            UncaughtExceptionHandler excHnd = new UncaughtExceptionHandler() {
-                @Override public void uncaughtException(Thread t, Throwable e) {
-                    if (grid != null)
-                        grid.context().failure().process(new FailureContext(FailureType.CRITICAL_ERROR, e));
-                }
-            };
-
-            execSvc = new IgniteThreadPoolExecutor(
-                "pub",
-                cfg.getIgniteInstanceName(),
-                cfg.getPublicThreadPoolSize(),
-                cfg.getPublicThreadPoolSize(),
-                DFLT_THREAD_KEEP_ALIVE_TIME,
-                new LinkedBlockingQueue<Runnable>(),
-                GridIoPolicy.PUBLIC_POOL,
-                oomeHnd);
-
-            execSvc.allowCoreThreadTimeOut(true);
-
-            validateThreadPoolSize(cfg.getServiceThreadPoolSize(), "service");
-
-            svcExecSvc = new IgniteThreadPoolExecutor(
-                "svc",
-                cfg.getGridName(),
-                cfg.getServiceThreadPoolSize(),
-                cfg.getServiceThreadPoolSize(),
-                DFLT_THREAD_KEEP_ALIVE_TIME,
-                new LinkedBlockingQueue<Runnable>(),
-                GridIoPolicy.SERVICE_POOL,
-                oomeHnd);
-
-            svcExecSvc.allowCoreThreadTimeOut(true);
-
-            validateThreadPoolSize(cfg.getSystemThreadPoolSize(), "system");
-
-            sysExecSvc = new IgniteThreadPoolExecutor(
-                "sys",
-                cfg.getIgniteInstanceName(),
-                cfg.getSystemThreadPoolSize(),
-                cfg.getSystemThreadPoolSize(),
-                DFLT_THREAD_KEEP_ALIVE_TIME,
-                new LinkedBlockingQueue<Runnable>(),
-                GridIoPolicy.SYSTEM_POOL,
-                oomeHnd);
-
-            sysExecSvc.allowCoreThreadTimeOut(true);
-
-            validateThreadPoolSize(cfg.getStripedPoolSize(), "stripedPool");
 
             WorkersRegistry workerRegistry = new WorkersRegistry(
                 new IgniteBiInClosure<GridWorker, FailureType>() {
@@ -1902,225 +1782,6 @@ public class IgnitionEx {
                     : cfg.getFailureDetectionTimeout()),
                 log);
 
-            stripedExecSvc = new StripedExecutor(
-                cfg.getStripedPoolSize(),
-                cfg.getIgniteInstanceName(),
-                "sys",
-                log,
-                new IgniteInClosure<Throwable>() {
-                    @Override public void apply(Throwable t) {
-                        if (grid != null)
-                            grid.context().failure().process(new FailureContext(SYSTEM_WORKER_TERMINATION, t));
-                    }
-                },
-                workerRegistry,
-                cfg.getFailureDetectionTimeout());
-
-            // Note that since we use 'LinkedBlockingQueue', number of
-            // maximum threads has no effect.
-            // Note, that we do not pre-start threads here as management pool may
-            // not be needed.
-            validateThreadPoolSize(cfg.getManagementThreadPoolSize(), "management");
-
-            mgmtExecSvc = new IgniteThreadPoolExecutor(
-                "mgmt",
-                cfg.getIgniteInstanceName(),
-                cfg.getManagementThreadPoolSize(),
-                cfg.getManagementThreadPoolSize(),
-                DFLT_THREAD_KEEP_ALIVE_TIME,
-                new LinkedBlockingQueue<Runnable>(),
-                GridIoPolicy.MANAGEMENT_POOL,
-                oomeHnd);
-
-            mgmtExecSvc.allowCoreThreadTimeOut(true);
-
-            // Note that since we use 'LinkedBlockingQueue', number of
-            // maximum threads has no effect.
-            // Note, that we do not pre-start threads here as class loading pool may
-            // not be needed.
-            validateThreadPoolSize(cfg.getPeerClassLoadingThreadPoolSize(), "peer class loading");
-
-            p2pExecSvc = new IgniteThreadPoolExecutor(
-                "p2p",
-                cfg.getIgniteInstanceName(),
-                cfg.getPeerClassLoadingThreadPoolSize(),
-                cfg.getPeerClassLoadingThreadPoolSize(),
-                DFLT_THREAD_KEEP_ALIVE_TIME,
-                new LinkedBlockingQueue<Runnable>(),
-                GridIoPolicy.P2P_POOL,
-                oomeHnd);
-
-            p2pExecSvc.allowCoreThreadTimeOut(true);
-
-            dataStreamerExecSvc = new StripedExecutor(
-                cfg.getDataStreamerThreadPoolSize(),
-                cfg.getIgniteInstanceName(),
-                "data-streamer",
-                log,
-                new IgniteInClosure<Throwable>() {
-                    @Override public void apply(Throwable t) {
-                        if (grid != null)
-                            grid.context().failure().process(new FailureContext(SYSTEM_WORKER_TERMINATION, t));
-                    }
-                },
-                true,
-                workerRegistry,
-                cfg.getFailureDetectionTimeout());
-
-            // Note that we do not pre-start threads here as this pool may not be needed.
-            validateThreadPoolSize(cfg.getAsyncCallbackPoolSize(), "async callback");
-
-            callbackExecSvc = new IgniteStripedThreadPoolExecutor(
-                cfg.getAsyncCallbackPoolSize(),
-                cfg.getIgniteInstanceName(),
-                "callback",
-                oomeHnd,
-                false,
-                0);
-
-            if (cfg.getConnectorConfiguration() != null) {
-                validateThreadPoolSize(cfg.getConnectorConfiguration().getThreadPoolSize(), "connector");
-
-                restExecSvc = new IgniteThreadPoolExecutor(
-                    "rest",
-                    cfg.getIgniteInstanceName(),
-                    cfg.getConnectorConfiguration().getThreadPoolSize(),
-                    cfg.getConnectorConfiguration().getThreadPoolSize(),
-                    DFLT_THREAD_KEEP_ALIVE_TIME,
-                    new LinkedBlockingQueue<>(),
-                    GridIoPolicy.UNDEFINED,
-                    oomeHnd
-                );
-
-                restExecSvc.allowCoreThreadTimeOut(true);
-            }
-
-            validateThreadPoolSize(cfg.getUtilityCacheThreadPoolSize(), "utility cache");
-
-            utilityCacheExecSvc = new IgniteThreadPoolExecutor(
-                "utility",
-                cfg.getIgniteInstanceName(),
-                cfg.getUtilityCacheThreadPoolSize(),
-                cfg.getUtilityCacheThreadPoolSize(),
-                cfg.getUtilityCacheKeepAliveTime(),
-                new LinkedBlockingQueue<>(),
-                GridIoPolicy.UTILITY_CACHE_POOL,
-                oomeHnd);
-
-            utilityCacheExecSvc.allowCoreThreadTimeOut(true);
-
-            affExecSvc = new IgniteThreadPoolExecutor(
-                "aff",
-                cfg.getIgniteInstanceName(),
-                1,
-                1,
-                DFLT_THREAD_KEEP_ALIVE_TIME,
-                new LinkedBlockingQueue<>(),
-                GridIoPolicy.AFFINITY_POOL,
-                oomeHnd);
-
-            affExecSvc.allowCoreThreadTimeOut(true);
-
-            if (IgniteComponentType.INDEXING.inClassPath()) {
-                int cpus = Runtime.getRuntime().availableProcessors();
-
-                idxExecSvc = new IgniteThreadPoolExecutor(
-                    "idx",
-                    cfg.getIgniteInstanceName(),
-                    cpus,
-                    cpus * 2,
-                    3000L,
-                    new LinkedBlockingQueue<>(1000),
-                    GridIoPolicy.IDX_POOL,
-                    oomeHnd
-                );
-
-                int buildIdxThreadPoolSize = cfg.getBuildIndexThreadPoolSize();
-
-                validateThreadPoolSize(buildIdxThreadPoolSize, "build-idx");
-
-                buildIdxExecSvc = new IgniteThreadPoolExecutor(
-                    "build-idx-runner",
-                    cfg.getIgniteInstanceName(),
-                    buildIdxThreadPoolSize,
-                    buildIdxThreadPoolSize,
-                    DFLT_THREAD_KEEP_ALIVE_TIME,
-                    new LinkedBlockingQueue<>(),
-                    GridIoPolicy.UNDEFINED,
-                    oomeHnd
-                );
-
-                buildIdxExecSvc.allowCoreThreadTimeOut(true);
-            }
-
-            validateThreadPoolSize(cfg.getQueryThreadPoolSize(), "query");
-
-            qryExecSvc = new IgniteThreadPoolExecutor(
-                "query",
-                cfg.getIgniteInstanceName(),
-                cfg.getQueryThreadPoolSize(),
-                cfg.getQueryThreadPoolSize(),
-                DFLT_THREAD_KEEP_ALIVE_TIME,
-                new LinkedBlockingQueue<>(),
-                GridIoPolicy.QUERY_POOL,
-                oomeHnd);
-
-            qryExecSvc.allowCoreThreadTimeOut(true);
-
-            schemaExecSvc = new IgniteThreadPoolExecutor(
-                "schema",
-                cfg.getIgniteInstanceName(),
-                2,
-                2,
-                DFLT_THREAD_KEEP_ALIVE_TIME,
-                new LinkedBlockingQueue<>(),
-                GridIoPolicy.SCHEMA_POOL,
-                oomeHnd);
-
-            schemaExecSvc.allowCoreThreadTimeOut(true);
-
-            validateThreadPoolSize(cfg.getRebalanceThreadPoolSize(), "rebalance");
-
-            rebalanceExecSvc = new IgniteThreadPoolExecutor(
-                "rebalance",
-                cfg.getIgniteInstanceName(),
-                cfg.getRebalanceThreadPoolSize(),
-                cfg.getRebalanceThreadPoolSize(),
-                DFLT_THREAD_KEEP_ALIVE_TIME,
-                new LinkedBlockingQueue<>(),
-                GridIoPolicy.UNDEFINED,
-                excHnd);
-
-            rebalanceExecSvc.allowCoreThreadTimeOut(true);
-
-            rebalanceStripedExecSvc = new IgniteStripedThreadPoolExecutor(
-                cfg.getRebalanceThreadPoolSize(),
-                cfg.getIgniteInstanceName(),
-                "rebalance-striped",
-                excHnd,
-                true,
-                DFLT_THREAD_KEEP_ALIVE_TIME);
-
-            if (!F.isEmpty(cfg.getExecutorConfiguration())) {
-                validateCustomExecutorsConfiguration(cfg.getExecutorConfiguration());
-
-                customExecSvcs = new HashMap<>();
-
-                for (ExecutorConfiguration execCfg : cfg.getExecutorConfiguration()) {
-                    ThreadPoolExecutor exec = new IgniteThreadPoolExecutor(
-                        execCfg.getName(),
-                        cfg.getIgniteInstanceName(),
-                        execCfg.getSize(),
-                        execCfg.getSize(),
-                        DFLT_THREAD_KEEP_ALIVE_TIME,
-                        new LinkedBlockingQueue<>(),
-                        GridIoPolicy.UNDEFINED,
-                        oomeHnd);
-
-                    customExecSvcs.put(execCfg.getName(), exec);
-                }
-            }
-
             // Register Ignite MBean for current grid instance.
             registerFactoryMbean(cfg.getMBeanServer());
 
@@ -2136,24 +1797,6 @@ public class IgnitionEx {
 
                 grid0.start(
                     cfg,
-                    utilityCacheExecSvc,
-                    execSvc,
-                    svcExecSvc,
-                    sysExecSvc,
-                    stripedExecSvc,
-                    p2pExecSvc,
-                    mgmtExecSvc,
-                    dataStreamerExecSvc,
-                    restExecSvc,
-                    affExecSvc,
-                    idxExecSvc,
-                    buildIdxExecSvc,
-                    callbackExecSvc,
-                    qryExecSvc,
-                    schemaExecSvc,
-                    rebalanceExecSvc,
-                    rebalanceStripedExecSvc,
-                    customExecSvcs,
                     new CA() {
                         @Override public void apply() {
                             startLatch.countDown();
@@ -2218,43 +1861,6 @@ public class IgnitionEx {
                 if (log.isDebugEnabled())
                     log.debug("Shutdown hook has not been installed because environment " +
                         "or system property " + IGNITE_NO_SHUTDOWN_HOOK + " is set.");
-            }
-        }
-
-        /**
-         * @param poolSize an actual value in the configuration.
-         * @param poolName a name of the pool like 'management'.
-         * @throws IgniteCheckedException If the poolSize is wrong.
-         */
-        private static void validateThreadPoolSize(int poolSize, String poolName)
-            throws IgniteCheckedException {
-            if (poolSize <= 0) {
-                throw new IgniteCheckedException("Invalid " + poolName + " thread pool size" +
-                    " (must be greater than 0), actual value: " + poolSize);
-            }
-        }
-
-        /**
-         * @param cfgs Array of the executors configurations.
-         * @throws IgniteCheckedException If configuration is wrong.
-         */
-        private static void validateCustomExecutorsConfiguration(ExecutorConfiguration[] cfgs)
-            throws IgniteCheckedException {
-            if (cfgs == null)
-                return;
-
-            Set<String> names = new HashSet<>(cfgs.length);
-
-            for (ExecutorConfiguration cfg : cfgs) {
-                if (F.isEmpty(cfg.getName()))
-                    throw new IgniteCheckedException("Custom executor name cannot be null or empty.");
-
-                if (!names.add(cfg.getName()))
-                    throw new IgniteCheckedException("Duplicate custom executor name: " + cfg.getName());
-
-                if (cfg.getSize() <= 0)
-                    throw new IgniteCheckedException("Custom executor size must be positive [name=" + cfg.getName() +
-                        ", size=" + cfg.getSize() + ']');
             }
         }
 
@@ -2792,9 +2398,6 @@ public class IgnitionEx {
 
                 grid = null;
 
-                if (log != null)
-                    stopExecutors(log);
-
                 log = null;
             }
         }
@@ -2878,108 +2481,6 @@ public class IgnitionEx {
             }
             finally {
                 grpCtx.preloader().resume();
-            }
-        }
-
-        /**
-         * Stops executor services if they has been started.
-         *
-         * @param log Grid logger.
-         */
-        private void stopExecutors(IgniteLogger log) {
-            boolean interrupted = Thread.interrupted();
-
-            try {
-                stopExecutors0(log);
-            }
-            finally {
-                if (interrupted)
-                    Thread.currentThread().interrupt();
-            }
-        }
-
-        /**
-         * Stops executor services if they has been started.
-         *
-         * @param log Grid logger.
-         */
-        private void stopExecutors0(IgniteLogger log) {
-            assert log != null;
-
-            U.shutdownNow(getClass(), execSvc, log);
-
-            execSvc = null;
-
-            U.shutdownNow(getClass(), svcExecSvc, log);
-
-            svcExecSvc = null;
-
-            U.shutdownNow(getClass(), sysExecSvc, log);
-
-            sysExecSvc = null;
-
-            U.shutdownNow(getClass(), qryExecSvc, log);
-
-            qryExecSvc = null;
-
-            U.shutdownNow(getClass(), schemaExecSvc, log);
-
-            schemaExecSvc = null;
-
-            U.shutdownNow(getClass(), rebalanceExecSvc, log);
-
-            rebalanceExecSvc = null;
-
-            U.shutdownNow(getClass(), rebalanceStripedExecSvc, log);
-
-            rebalanceStripedExecSvc = null;
-
-            U.shutdownNow(getClass(), stripedExecSvc, log);
-
-            stripedExecSvc = null;
-
-            U.shutdownNow(getClass(), mgmtExecSvc, log);
-
-            mgmtExecSvc = null;
-
-            U.shutdownNow(getClass(), p2pExecSvc, log);
-
-            p2pExecSvc = null;
-
-            U.shutdownNow(getClass(), dataStreamerExecSvc, log);
-
-            dataStreamerExecSvc = null;
-
-            if (restExecSvc != null)
-                U.shutdownNow(getClass(), restExecSvc, log);
-
-            restExecSvc = null;
-
-            U.shutdownNow(getClass(), utilityCacheExecSvc, log);
-
-            utilityCacheExecSvc = null;
-
-            U.shutdownNow(getClass(), affExecSvc, log);
-
-            affExecSvc = null;
-
-            U.shutdownNow(getClass(), idxExecSvc, log);
-
-            idxExecSvc = null;
-
-            U.shutdownNow(getClass(), buildIdxExecSvc, log);
-
-            buildIdxExecSvc = null;
-
-            U.shutdownNow(getClass(), callbackExecSvc, log);
-
-            callbackExecSvc = null;
-
-            if (!F.isEmpty(customExecSvcs)) {
-                for (ThreadPoolExecutor exec : customExecSvcs.values())
-                    U.shutdownNow(getClass(), exec, log);
-
-                customExecSvcs = null;
             }
         }
