+++ modules/core/src/main/java/org/apache/ignite/internal/managers/encryption/GridEncryptionManager.java
@@ -927,7 +927,7 @@ public class GridEncryptionManager extends GridManagerAdapter<EncryptionSpi> imp
      */
     public void onWalSegmentRemoved(long segmentIdx) {
         if (grpKeys.isReleaseWalKeysRequired(segmentIdx))
-            ctx.getSystemExecutorService().submit(() -> releaseWalKeys(segmentIdx));
+            ctx.pools().getSystemExecutorService().submit(() -> releaseWalKeys(segmentIdx));
     }
 
     /**
