+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/distributed/dht/topology/PartitionsEvictManager.java
@@ -210,7 +210,7 @@ public class PartitionsEvictManager extends GridCacheSharedManagerAdapter {
     @Override protected void start0() throws IgniteCheckedException {
         super.start0();
 
-        executor = (IgniteThreadPoolExecutor) cctx.kernalContext().getRebalanceExecutorService();
+        executor = (IgniteThreadPoolExecutor) cctx.kernalContext().pools().getRebalanceExecutorService();
     }
 
     /** {@inheritDoc} */
