+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/mvcc/MvccProcessorImpl.java
@@ -602,7 +602,7 @@ public class MvccProcessorImpl extends GridProcessorAdapter implements MvccProce
             prevQueries.init(nodes, ctx.discovery()::alive);
         }
         else if (sndQrys) {
-            ctx.getSystemExecutorService().submit(() -> {
+            ctx.pools().getSystemExecutorService().submit(() -> {
                 try {
                     sendMessage(newCrd.nodeId(), new MvccActiveQueriesMessage(qryIds));
                 }
