+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/GridCacheDatabaseSharedManager.java
@@ -1414,7 +1414,7 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
         if (fut.localJoinExchange() || fut.activateCluster()
             || (fut.exchangeActions() != null && !F.isEmpty(fut.exchangeActions().cacheGroupsToStart()))) {
             U.doInParallel(
-                cctx.kernalContext().getSystemExecutorService(),
+                cctx.kernalContext().pools().getSystemExecutorService(),
                 cctx.cache().cacheGroups(),
                 cacheGroup -> {
                     if (cacheGroup.isLocal())
@@ -2150,7 +2150,7 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
 
         AtomicReference<Throwable> applyError = new AtomicReference<>();
 
-        StripedExecutor exec = cctx.kernalContext().getStripedExecutorService();
+        StripedExecutor exec = cctx.kernalContext().pools().getStripedExecutorService();
 
         Semaphore semaphore = new Semaphore(semaphorePertmits(exec));
 
@@ -2642,7 +2642,7 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
 
         long lastArchivedSegment = cctx.wal().lastArchivedSegment();
 
-        StripedExecutor exec = cctx.kernalContext().getStripedExecutorService();
+        StripedExecutor exec = cctx.kernalContext().pools().getStripedExecutorService();
 
         Semaphore semaphore = new Semaphore(semaphorePertmits(exec));
