+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/snapshot/SnapshotRestoreProcess.java
@@ -373,7 +373,7 @@ public class SnapshotRestoreProcess {
             if (fut0 != null && reqId.equals(fut0.rqId)) {
                 fut = null;
 
-                ctx.getSystemExecutorService().submit(() -> fut0.onDone(null, err));
+                ctx.pools().getSystemExecutorService().submit(() -> fut0.onDone(null, err));
             }
         }
     }
