+++ modules/core/src/main/java/org/apache/ignite/internal/processors/rest/GridRestProcessor.java
@@ -210,7 +210,7 @@ public class GridRestProcessor extends GridProcessorAdapter implements IgniteRes
             fut.setWorker(w);
 
             try {
-                ctx.getRestExecutorService().execute(w);
+                ctx.pools().getRestExecutorService().execute(w);
             }
             catch (RejectedExecutionException e) {
                 U.error(log, "Failed to execute worker due to execution rejection " +
