+++ modules/core/src/test/java/org/apache/ignite/internal/util/future/GridFutureAdapterSelfTest.java
@@ -179,10 +179,20 @@ public class GridFutureAdapterSelfTest extends GridCommonAbstractTest {
     public void testListenNotify() throws Exception {
         GridTestKernalContext ctx = new GridTestKernalContext(log);
 
-        ctx.setExecutorService(Executors.newSingleThreadExecutor(new IgniteThreadFactory("testscope", "exec-svc")));
-        ctx.setSystemExecutorService(Executors.newSingleThreadExecutor(new IgniteThreadFactory("testscope", "system-exec")));
+        ctx.add(new PoolProcessor(ctx) {
+            final ExecutorService execSvc = Executors.newSingleThreadExecutor(new IgniteThreadFactory("testscope", "exec-svc"));
 
-        ctx.add(new PoolProcessor(ctx));
+            final ExecutorService sysExecSvc = Executors.newSingleThreadExecutor(new IgniteThreadFactory("testscope", "system-exec"));
+
+            @Override public ExecutorService getSystemExecutorService() {
+                return sysExecSvc;
+            }
+
+            @Override public ExecutorService getExecutorService() {
+                return execSvc;
+            }
+        });
+
         ctx.add(new GridClosureProcessor(ctx));
 
         ctx.start();
