+++ modules/core/src/test/java/org/apache/ignite/testframework/junits/GridTestKernalContext.java
@@ -21,7 +21,6 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.ListIterator;
-import java.util.concurrent.ExecutorService;
 import org.apache.ignite.IgniteCheckedException;
 import org.apache.ignite.IgniteLogger;
 import org.apache.ignite.configuration.IgniteConfiguration;
@@ -68,24 +67,6 @@ public class GridTestKernalContext extends GridKernalContextImpl {
                 new IgniteKernal(null),
                 cfg,
                 new GridKernalGatewayImpl(null),
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
                 cfg.getPluginProviders() != null && cfg.getPluginProviders().length > 0 ?
                     Arrays.asList(cfg.getPluginProviders()) : U.allPluginProviders(),
                 null,
@@ -131,24 +112,6 @@ public class GridTestKernalContext extends GridKernalContextImpl {
 
             comp.stop(cancel);
         }
-    }
-
-    /**
-     * Sets system executor service.
-     *
-     * @param sysExecSvc Executor service
-     */
-    public void setSystemExecutorService(ExecutorService sysExecSvc) {
-        this.sysExecSvc = sysExecSvc;
-    }
-
-    /**
-     * Sets executor service.
-     *
-     * @param execSvc Executor service
-     */
-    public void setExecutorService(ExecutorService execSvc) {
-        this.execSvc = execSvc;
     }
 
     /** {@inheritDoc} */
